refactor(Modularity): Separate swagger configuration logic
#22
